requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
gevent==23.9.1
pyautogui==0.9.54
qrcode==7.4.2
pillow==10.3.0
pystray==0.19.5